Combat fixes

+++ Elder_Shadows/Assets/Scripts/Inventory/IModifiers.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IModifier
 {
-    void AddValue(ref int baseValue);
+    void AddValue(ref float baseValue);
 }
